Refactor map speed handling and update state machine wait time (#32)
- Remove unused `wmap_getMapScaledSpeed` function and its references
- Simplify `wmap_getMapScaledTime` function to directly use `globalSpeed`
- Update projectile speed calculations to use raw speed values
- Change `waitEndGameTime` from `f32` to `f64` for better precision
- Adjust wait state creation to directly calculate `waitEndGameTime`

+++ src/war_state_machine_wait.c
@@ -5,7 +5,7 @@
 WarState* wst_createWaitState(WarContext* context, WarEntity* entity, f32 waitTime)
 {
     WarState* state = wst_createState(context, entity, WAR_STATE_WAIT);
-    state->wait.waitEndGameTime = (f32)(context->gameTime + waitTime);
+    state->wait.waitEndGameTime = context->gameTime + waitTime;
     return state;
 }
